hw4/pyproject.toml
[tool.poetry]
name = "workspace"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
numpy = "1.23.5"
tensorflow-cpu = "^2.11.0"
opencv-python = "^4.7.0.72"
scipy = "^1.10.1"
pillow = "^9.4.0"
scikit-image = "^0.19.3"
black = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
